+++ src/Entity.cpp
@@ -20,11 +20,11 @@ void Entity::setOrigin(const sf::Vector2f& origin) {
 
 bool Entity::isAligned(const char axis) const {
 	switch (axis) {
-		case 'x': return (unsigned short)position.x % Game::TILE_SIZE == 0;
-		case 'y': return (unsigned short)position.y % Game::TILE_SIZE == 0;
+		case 'x': return int(position.x) % Game::TILE_SIZE == 0;
+		case 'y': return int(position.y) % Game::TILE_SIZE == 0;
 	}
-	return (unsigned short)position.x % Game::TILE_SIZE == 0 
-		&& (unsigned short)position.y % Game::TILE_SIZE == 0;
+	return int(position.x) % Game::TILE_SIZE == 0 
+		&& int(position.y) % Game::TILE_SIZE == 0;
 }
 
 void Entity::update() {
